Refactors webpack module configuration module
 - Loaders are defined in _loaders (previouly webpack-loader-mixin)
 - Loader utils for style loaders are in _style-utils
 - Moved webpack-module config to module/
