+++ webpack/config/module/_loaders.js
@@ -1,95 +1,5 @@
-const MiniCssExtractPlugin = require('mini-css-extract-plugin');
-
-const webpackEnvModule = require('./webpack-env-module');
-const paths = require('../utils/paths');
-
-const styleTypes = {
-  css: {
-    test: /\.css$/,
-    exclude: /\.module\.css$/,
-    // Don't consider CSS imports dead code even if the
-    // containing package claims to have no side effects.
-    // Remove this when webpack adds a warning or an error for this.
-    // See https://github.com/webpack/webpack/issues/6571
-    sideEffects: true,
-  },
-  cssModules: {
-    test: /\.css$/,
-  },
-  // TODO define SCSS options
-  scss: true,
-};
-
-// common function to get style loaders
-const getStyleLoaders = ({ cssLoaderOptions, preprocessors }) => {
-  const isEnvDevelopment = webpackEnvModule.isEnvDevelopment();
-  const isEnvProduction = webpackEnvModule.isEnvProduction();
-  const shouldUseSourceMap = webpackEnvModule.shouldUseSourceMap();
-  const shouldUseRelativeAssetPaths = webpackEnvModule.shouldUseRelativeAssetPaths();
-
-  // Defines 'use' property for style loaders.
-  const loaders = [
-    isEnvDevelopment && require.resolve('vue-style-loader'),
-    isEnvProduction && {
-      loader: MiniCssExtractPlugin.loader,
-      options: shouldUseRelativeAssetPaths
-        ? { publicPath: '../../' }
-        : {},
-    },
-    {
-      loader: require.resolve('css-loader'),
-      options: cssLoaderOptions,
-    },
-    {
-      // Options for PostCSS as we reference these options twice
-      // Adds vendor prefixing based on your specified browser support in
-      // package.json
-      loader: require.resolve('postcss-loader'),
-      options: {
-        // Necessary for external CSS imports to work
-        // https://github.com/facebook/create-react-app/issues/2677
-        ident: 'postcss',
-        plugins: () => [
-          require('postcss-flexbugs-fixes'),
-          require('postcss-preset-env')({
-            autoprefixer: {
-              flexbox: 'no-2009',
-            },
-            stage: 3,
-          }),
-        ],
-        sourceMap: isEnvProduction && shouldUseSourceMap,
-      },
-    },
-  ].filter(Boolean);
-
-  if (preprocessors && preprocessors.length > 0) {
-    // TODO verify its importance.
-    // Handles url() calls for SCSS
-    // loaders.push({
-    //   loader: require.resolve('resolve-url-loader'),
-    //   options: {
-    //     sourceMap: isEnvProduction && shouldUseSourceMap,
-    //   },
-    // });
-
-    preprocessors.forEach(({ name, options }) => {
-      loaders.push(
-        {
-          loader: require.resolve(name),
-          options: Object.assign(
-            {},
-            {
-              sourceMap: true,
-            },
-            options,
-          ),
-        },
-      );
-    });
-  }
-  return loaders;
-};
+const stylesUtil = require('./_styles-util');
+const paths = require('../../utils/paths');
 
 module.exports = {
   configureBabelLoader() {
@@ -123,14 +33,14 @@ module.exports = {
   },
 
   configureStyleLoader(args) {
-    if (!(args.type in styleTypes)) {
-      throw new TypeError('')
+    if (!(args.type in stylesUtil.styleTypes)) {
+      throw new TypeError('No style loader configuration for type ' + args.type, args);
     }
 
-    const styleLoaderOptions = styleTypes[args.type];
+    const styleLoaderOptions = stylesUtil.styleTypes[args.type];
     return {
       ...styleLoaderOptions,
-      use: getStyleLoaders(args),
+      use: stylesUtil.getStyleLoaders(args),
     };
   },
 
@@ -148,7 +58,7 @@ module.exports = {
       // by webpack's internal loaders.
       exclude: [/\.(js|mjs|vue)$/, /\.html$/, /\.json$/],
       options: {
-        name: 'static/media[name].[hash:8].[ext]',
+        name: 'static/media/[name].[hash:8].[ext]',
       },
     }
   },
