+++ webpack/config/module/webpack-module.config.js
@@ -1,5 +1,5 @@
-const configurator = require('../builder/webpack-configurator');
-const webpackEnvModule = require('../builder/webpack-env-module');
+const loaders = require('./_loaders');
+const webpackEnvModule = require('../../builder/webpack-env-module');
 
 const imageInlineSizeLimit = parseInt(
   process.env.IMAGE_INLINE_SIZE_LIMIT || '10000'
@@ -13,22 +13,22 @@ module.exports = () => {
     strictExportPresence: true,
     rules: [
       // { eslint }
-      configurator.configureVueLoader(),
+      loaders.configureVueLoader(),
       {
         // "oneOf" will travers all of the following loaders until one will match
         // the requirements. When no loader matches, it will fall back to the
         // "file" loader at the end of the loader list.
         oneOf: [
-          configurator.configureImageLoader(imageInlineSizeLimit),
-          configurator.configureMediaLoader(imageInlineSizeLimit),
-          configurator.configureFontLoader(imageInlineSizeLimit),
-          configurator.configureBabelLoader(),
+          loaders.configureImageLoader(imageInlineSizeLimit),
+          loaders.configureMediaLoader(imageInlineSizeLimit),
+          loaders.configureFontLoader(imageInlineSizeLimit),
+          loaders.configureBabelLoader(),
           {
             test: /\.css$/,
             oneOf: [
               {
                 resourceQuery: /module/,
-                ...configurator.configureStyleLoader({
+                ...loaders.configureStyleLoader({
                   type: 'cssModules',
                   cssLoaderOptions: {
                     importLoaders: 1,
@@ -42,7 +42,7 @@ module.exports = () => {
                   },
                 }),
               },
-              configurator.configureStyleLoader({
+              loaders.configureStyleLoader({
                 type: 'css',
                 cssLoaderOptions: {
                   importLoaders: 1,
@@ -51,7 +51,7 @@ module.exports = () => {
               }),
             ],
           },
-          configurator.configureFallbackLoader(),
+          loaders.configureFallbackLoader(),
         ],
       },
     ],
